修复保存项目信息的时候没有保存的发起日期和金额

+++ crowdfunding-member-parent/crowdfunding-member-entity/src/main/java/com/company/entity/vo/ProjectVO.java
@@ -26,7 +26,7 @@ public class ProjectVO implements Serializable {
 	private String projectDescription;
 	
 	// 计划筹集的金额
-	private Integer money;
+	private Long money;
 	
 	// 筹集资金的天数
 	private Integer day;
